recordController updated

+++ controllers/debtsController.js
@@ -1,4 +1,5 @@
 const DebtSchema = require("../models/Debt");
+const AccountSchema = require("../models/Account");
 
 exports.addDebt = async (req, res) => {
   const userId = req.userId;
@@ -11,6 +12,10 @@ exports.addDebt = async (req, res) => {
     if (isNaN(amount) || amount <= 0) {
       return res.status(400).json({ message: "Amount must be a positive number!" });
     }
+    const account = await AccountSchema.findById({ userId: req.userId, _id: accountId });
+    if( !account ) {
+      return res.status(400).json({ message: "Account not found!" });
+    }
 
     const debt = DebtSchema({
       type,
@@ -22,6 +27,16 @@ exports.addDebt = async (req, res) => {
       Date,
       DueDate
     });
+
+    if ( debt.type === "lend" ) {
+      account.balance = account.balance - amount;
+      await account.save();
+    }
+
+    if ( debt.type === "owe" ) {
+      account.balance = account.balance + amount;
+      await account.save();
+    }
 
     await debt.save();
     res.status(200).json({ message: "Debt Added" });
@@ -69,10 +84,25 @@ exports.patchDebt = async (req, res) => {
           return res.status(400).json({ message: "You cannot change the type of debt!" });
         }
         if (debt.accountId !== accountId) {
-          return res.status(400).json({ message: "You cannot change the finance account!" });
+          return res.status(400).json({ message: "You cannot change the financial account!" });
         }
         if (name) debt.name = name;
-        if (amount) debt.amount = amount;
+        if (amount) {
+          if (debt.type === "lend") {
+            const account = await AccountSchema.findOne({ userId: req.userId, _id: accountId });
+            account.balance = account.balance + debt.amount;
+            debt.amount = amount;
+            account.balance = account.balance - debt.amount;
+            await account.save();
+          }
+          if (debt.type === "owe") {
+            const account = await AccountSchema.findOne({ userId: req.userId, _id: accountId });
+            account.balance = account.balance - debt.amount;
+            debt.amount = amount;
+            account.balance = account.balance + debt.amount;
+            await account.save();
+          }
+        }
         if(purpose) debt.purpose = purpose;
         if (Date) debt.Date = Date;
         if (DueDate) debt.DueDate = DueDate;
@@ -87,7 +117,20 @@ exports.patchDebt = async (req, res) => {
 
 exports.deleteDebt = async (req, res) => {
   const { debtId } = req.params;
-  DebtSchema.findOneAndDelete({ userId: req.userId, _id: debtId})
+  const debt = await DebtSchema.findOne({ useId: req.userId, _id: debtId });
+  if (!debt) {
+    return res.status(404).json({ message: "Debt not found!" });
+  }
+  // const account = await AccountSchema.findOne({ userId: req.userId, _id: debt.accountId });
+  // if (debt.type == "lend") {
+  //   account.balance = account.balance + debt.amount;
+  //   await account.save();
+  // }
+  // if (debt.type == "owe") {
+  //   account.balance = account.balance - debt.amount;
+  //   await account.save();
+  // }
+  await DebtSchema.findOneAndDelete({ userId: req.userId, _id: debtId})
     .then(() => {
       res.status(200).json({ message: "Debt Deleted" });
     })
